fix(api): resolve current version to latest events
