+++ .github/workflows/check.yml
@@ -16,3 +16,5 @@ jobs:
           npm install
       - name: Lint
         run: npm run lint
+      - name: Test
+        run: npm test
