+++ package.json
@@ -6,6 +6,7 @@
     "dev": "next dev --webpack",
     "build": "next build --webpack",
     "start": "next start",
+    "test": "tsx --test \"src/**/*.test.ts\"",
     "lint": "eslint .",
     "format": "eslint . --fix && prettier --write ."
   },
@@ -34,6 +35,7 @@
     "prettier": "3.9.6",
     "serwist": "9.5.12",
     "tailwindcss": "4.3.3",
+    "tsx": "4.23.13",
     "typescript": "6.0.3"
   },
   "engines": {
